feat : 특정 요청시 디바이스토큰 받도록 변경

+++ src/main/java/com/depromeet/breadmapbackend/domain/user/UserServiceImpl.java
@@ -81,13 +81,13 @@ public class UserServiceImpl implements UserService {
 	public AlarmDto alarmChange(String oAuthId, NoticeTokenRequest request) {
 		User user = userRepository.findByOAuthId(oAuthId)
 			.orElseThrow(() -> new DaedongException(DaedongStatus.USER_NOT_FOUND));
-		if (!user.getIsAlarmOn()) { // 알림 off이면
+		if (request.isNoticeAgree()) {
 			if (noticeTokenRepository.findByUserAndDeviceToken(user, request.getDeviceToken()).isEmpty()) {
 				noticeTokenRepository.save(
 					NoticeToken.builder().user(user).deviceToken(request.getDeviceToken()).build());
 			}
 			return new AlarmDto(user.alarmOn());
-		} else { // 알림 on이면
+		} else {
 			if (noticeTokenRepository.findByUserAndDeviceToken(user, request.getDeviceToken()).isPresent()) {
 				noticeTokenRepository.delete(
 					noticeTokenRepository.findByUserAndDeviceToken(user, request.getDeviceToken()).get());
